chore(api-client): rename API files to api-client
